// sm_vue_client/src/openapi/system/models/BocSmIdentityOptionsDtosIdentityPasswordOptionsSettingDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BocSmIdentityOptionsDtosIdentityPasswordOptionsSettingDto = {
    requiredLength?: number;
    requiredUniqueChars?: number;
    requireNonAlphanumeric?: boolean;
    requireLowercase?: boolean;
    requireUppercase?: boolean;
    requireDigit?: boolean;
};